Refactoring to use new next-auth module
The authentication code that was part of this project has not been moved off into seperate modules, to simplify the code base and make it easier to work with.

The code is now included in `next-auth` and `next-auth-client` modules on npm. It’s faster, more elegant, contains additional bug fixes and addresses some corner-case bugs.

While the `next-auth` project includes it’s own sample implementation, this projects provides a more complete example.