Sign in dialog edge case improvement
If the call to /auth/providers fails and it returns a garbage response (e.g. due to a bad network connection) then fail gracefully and just don’t display the oAuth sign in buttons.

This is a silly edge case, as if the network is that bad you probably won’t be able to sign in, but it can happen and if not handled this way will display inactive buttons in invalid values (as it will attempt to iterate over the Error object!).