Client side fix to prevent duplicate signin emails
The sign in form now disables the submit button and input field (and shows a loading spinner) while it is submitting an email sign in request.

This is helpful as sometimes the form takes a second or two to submit on the demo site so avoids multiple emails being generated.

This ideally should be fixed server side, but this is a quicker enhancement.