Removing `now-logs` dependancy
The `now-logs` package relies on an outdated version of socket.io with known high severity security issues flagged by `npm audit`.

The package has not been updated in a year. I created a patch for it to address this issue but the service that invokes it is currently unavailable as there are billing restrictions on it so have decided just to remove `now-logs` from the project.